# walletnode/api/gen/http/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "WalletNode REST API",
    "version": "1.0"
  },
  "host": "localhost:8080",
  "consumes": [
    "application/json",
    "application/xml",
    "application/gob"
  ],
  "produces": [
    "application/json",
    "application/xml",
    "application/gob"
  ],
  "paths": {
    "/artworks/register": {
      "get": {
        "tags": [
          "artworks"
        ],
        "summary": "Returns list of tasks",
        "description": "List of all tasks.",
        "operationId": "artworks#registerTasks",
        "responses": {
          "200": {
            "description": "OK response.",
            "schema": {
              "$ref": "#/definitions/ArtworksTaskResponseTinyCollection"
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterTasksInternalServerErrorResponseBody"
            }
          }
        },
        "schemes": [
          "http"
        ]
      },
      "post": {
        "tags": [
          "artworks"
        ],
        "summary": "Registers a new artwork",
        "description": "Runs a new registration process for the new artwork.",
        "operationId": "artworks#register",
        "parameters": [
          {
            "name": "RegisterRequestBody",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterRequestBody",
              "required": [
                "image_id",
                "artist_name",
                "name",
                "issued_copies",
                "artist_pastelid",
                "artist_pastelid_passphrase",
                "spendable_address",
                "maximum_fee"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterResponseBody"
            }
          },
          "400": {
            "description": "Bad Request response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterBadRequestResponseBody"
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterInternalServerErrorResponseBody"
            }
          }
        },
        "schemes": [
          "http"
        ]
      }
    },
    "/artworks/register/upload": {
      "post": {
        "tags": [
          "artworks"
        ],
        "summary": "Uploads an image",
        "description": "Upload the image that is used when registering a new artwork.",
        "operationId": "artworks#uploadImage",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "UploadImageRequestBody",
            "in": "formData",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ArtworksUploadImageRequestBody",
              "required": [
                "file"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created response.",
            "schema": {
              "$ref": "#/definitions/ArtworksUploadImageResponseBody"
            }
          },
          "400": {
            "description": "Bad Request response.",
            "schema": {
              "$ref": "#/definitions/ArtworksUploadImageBadRequestResponseBody"
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "schema": {
              "$ref": "#/definitions/ArtworksUploadImageInternalServerErrorResponseBody"
            }
          }
        },
        "schemes": [
          "http"
        ]
      }
    },
    "/artworks/register/{taskId}": {
      "get": {
        "tags": [
          "artworks"
        ],
        "summary": "Find task by ID",
        "description": "Returns a single task.",
        "operationId": "artworks#registerTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID of the registration process",
            "required": true,
            "type": "string",
            "maxLength": 8,
            "minLength": 8
          }
        ],
        "responses": {
          "200": {
            "description": "OK response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterTaskResponseBody"
            }
          },
          "404": {
            "description": "Not Found response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterTaskNotFoundResponseBody"
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterTaskInternalServerErrorResponseBody"
            }
          }
        },
        "schemes": [
          "http"
        ]
      }
    },
    "/artworks/register/{taskId}/state": {
      "get": {
        "tags": [
          "artworks"
        ],
        "summary": "Streams state by task ID",
        "description": "Streams the state of the registration process.",
        "operationId": "artworks#registerTaskState",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "Task ID of the registration process",
            "required": true,
            "type": "string",
            "maxLength": 8,
            "minLength": 8
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterTaskStateResponseBody",
              "required": [
                "date",
                "status"
              ]
            }
          },
          "404": {
            "description": "Not Found response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterTaskStateNotFoundResponseBody"
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "schema": {
              "$ref": "#/definitions/ArtworksRegisterTaskStateInternalServerErrorResponseBody"
            }
          }
        },
        "schemes": [
          "ws"
        ]
      }
    },
    "/artworks/search/{term}": {
      "get": {
        "tags": [
          "artworks"
        ],
        "summary": "search artworks",
        "description": "Search artwork tickets",
        "operationId": "artworks#searchRequest",
        "parameters": [
          {
            "name": "term",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "json"
          }
        ],
        "responses": {
          "101": {
            "description": "Switching Protocols response.",
            "schema": {
              "$ref": "#/definitions/ArtworksSearchRequestResponseBody"
            }
          },
          "400": {
            "description": "Bad Request response.",
            "schema": {
              "$ref": "#/definitions/ArtworksSearchRequestBadRequestResponseBody"
            }
          },
          "500": {
            "description": "Internal Server Error response.",
            "schema": {
              "$ref": "#/definitions/ArtworksSearchRequestInternalServerErrorResponseBody"
            }
          }
        },
        "schemes": [
          "ws"
        ]
      }
    }
  },
  "definitions": {
...
    "ArtworksRegisterBadRequestResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": false
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": true
        }
      },
      "description": "register_BadRequest_response_body result type (default view)",
      "example": {
        "fault": false,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": false
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
    "ArtworksRegisterInternalServerErrorResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": false
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": true
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": true
        }
      },
      "description": "register_InternalServerError_response_body result type (default view)",
      "example": {
        "fault": true,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": true
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
...
    "ArtworksRegisterResponseBody": {
      "title": "Mediatype identifier: application/vnd.artwork.register; view=default",
      "type": "object",
      "properties": {
        "task_id": {
          "type": "string",
          "description": "Task ID of the registration process",
          "example": "n6Qn6TFM",
          "minLength": 8,
          "maxLength": 8
        }
      },
      "description": "RegisterResponseBody result type (default view)",
      "example": {
        "task_id": "n6Qn6TFM"
      },
      "required": [
        "task_id"
      ]
    },
...
    "ArtworksRegisterTasksInternalServerErrorResponseBody": {
      "title": "Mediatype identifier: application/vnd.goa.error; view=default",
      "type": "object",
      "properties": {
        "fault": {
          "type": "boolean",
          "description": "Is the error a server-side fault?",
          "example": true
        },
        "id": {
          "type": "string",
          "description": "ID is a unique identifier for this particular occurrence of the problem.",
          "example": "123abc"
        },
        "message": {
          "type": "string",
          "description": "Message is a human-readable explanation specific to this occurrence of the problem.",
          "example": "parameter 'p' must be an integer"
        },
        "name": {
          "type": "string",
          "description": "Name is the name of this class of errors.",
          "example": "bad_request"
        },
        "temporary": {
          "type": "boolean",
          "description": "Is the error temporary?",
          "example": true
        },
        "timeout": {
          "type": "boolean",
          "description": "Is the error a timeout?",
          "example": true
        }
      },
      "description": "registerTasks_InternalServerError_response_body result type (default view)",
      "example": {
        "fault": true,
        "id": "123abc",
        "message": "parameter 'p' must be an integer",
        "name": "bad_request",
        "temporary": false,
        "timeout": true
      },
      "required": [
        "name",
        "id",
        "message",
        "temporary",
        "timeout",
        "fault"
      ]
    },
...
    "ArtworksTaskResponseTinyCollection": {
      "title": "Mediatype identifier: application/vnd.artwork.register.task; type=collection; view=tiny",
      "type": "array",
      "items": {
        "$ref": "#/definitions/TaskResponseTiny"
      },
      "description": "ArtworksTaskResponseTinyCollection is the result type for an array of TaskResponseTiny (default view)",
      "example": [
        {
          "id": "n6Qn6TFM",
          "states": [
            {
              "date": "2006-01-02T15:04:05Z07:00",
              "status": "Task Started"
            },
            {
              "date": "2006-01-02T15:04:05Z07:00",
              "status": "Task Started"
            }
          ],
          "status": "Task Started",
          "ticket": {
            "artist_name": "Leonardo da Vinci",
            "artist_pastelid": "jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS",
            "artist_pastelid_passphrase": "qwerasdf1234",
            "artist_website_url": "https://www.leonardodavinci.net",
            "description": "The Mona Lisa is an oil painting by Italian artist, inventor, and writer Leonardo da Vinci. Likely completed in 1506, the piece features a portrait of a seated woman set against an imaginary landscape.",
            "issued_copies": 1,
            "keywords": "Renaissance, sfumato, portrait",
            "maximum_fee": 100,
            "name": "Mona Lisa",
            "series_name": "Famous artist",
            "spendable_address": "PtiqRXn2VQwBjp1K8QXR2uW2w2oZ3Ns7N6j",
            "youtube_url": "https://www.youtube.com/watch?v=0xl6Ufo4ZX0"
          },
          "txid": "576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58"
        },
        {
          "id": "n6Qn6TFM",
          "states": [
            {
              "date": "2006-01-02T15:04:05Z07:00",
              "status": "Task Started"
            },
            {
              "date": "2006-01-02T15:04:05Z07:00",
              "status": "Task Started"
            }
          ],
          "status": "Task Started",
          "ticket": {
            "artist_name": "Leonardo da Vinci",
            "artist_pastelid": "jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS",
            "artist_pastelid_passphrase": "qwerasdf1234",
            "artist_website_url": "https://www.leonardodavinci.net",
            "description": "The Mona Lisa is an oil painting by Italian artist, inventor, and writer Leonardo da Vinci. Likely completed in 1506, the piece features a portrait of a seated woman set against an imaginary landscape.",
            "issued_copies": 1,
            "keywords": "Renaissance, sfumato, portrait",
            "maximum_fee": 100,
            "name": "Mona Lisa",
            "series_name": "Famous artist",
            "spendable_address": "PtiqRXn2VQwBjp1K8QXR2uW2w2oZ3Ns7N6j",
            "youtube_url": "https://www.youtube.com/watch?v=0xl6Ufo4ZX0"
          },
          "txid": "576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58"
        },
        {
          "id": "n6Qn6TFM",
          "states": [
            {
              "date": "2006-01-02T15:04:05Z07:00",
              "status": "Task Started"
            },
            {
              "date": "2006-01-02T15:04:05Z07:00",
              "status": "Task Started"
            }
          ],
          "status": "Task Started",
          "ticket": {
            "artist_name": "Leonardo da Vinci",
            "artist_pastelid": "jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS",
            "artist_pastelid_passphrase": "qwerasdf1234",
            "artist_website_url": "https://www.leonardodavinci.net",
            "description": "The Mona Lisa is an oil painting by Italian artist, inventor, and writer Leonardo da Vinci. Likely completed in 1506, the piece features a portrait of a seated woman set against an imaginary landscape.",
            "issued_copies": 1,
            "keywords": "Renaissance, sfumato, portrait",
            "maximum_fee": 100,
            "name": "Mona Lisa",
            "series_name": "Famous artist",
            "spendable_address": "PtiqRXn2VQwBjp1K8QXR2uW2w2oZ3Ns7N6j",
            "youtube_url": "https://www.youtube.com/watch?v=0xl6Ufo4ZX0"
          },
          "txid": "576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58"
        },
        {
          "id": "n6Qn6TFM",
          "states": [
            {
              "date": "2006-01-02T15:04:05Z07:00",
              "status": "Task Started"
            },
            {
              "date": "2006-01-02T15:04:05Z07:00",
              "status": "Task Started"
            }
          ],
          "status": "Task Started",
          "ticket": {
            "artist_name": "Leonardo da Vinci",
            "artist_pastelid": "jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS",
            "artist_pastelid_passphrase": "qwerasdf1234",
            "artist_website_url": "https://www.leonardodavinci.net",
            "description": "The Mona Lisa is an oil painting by Italian artist, inventor, and writer Leonardo da Vinci. Likely completed in 1506, the piece features a portrait of a seated woman set against an imaginary landscape.",
            "issued_copies": 1,
            "keywords": "Renaissance, sfumato, portrait",
            "maximum_fee": 100,
            "name": "Mona Lisa",
            "series_name": "Famous artist",
            "spendable_address": "PtiqRXn2VQwBjp1K8QXR2uW2w2oZ3Ns7N6j",
            "youtube_url": "https://www.youtube.com/watch?v=0xl6Ufo4ZX0"
          },
          "txid": "576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58"
        }
      ]
    },
...
  }
}